backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.32.0
Django==4.1.7
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework==3.14.0
dropbox==11.36.0
gcloud==0.17.0
google-api-core==2.11.0
google-auth==2.16.1
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
jws==0.1.3
oauth2client==3.0.0
Pillow==9.4.0
ply==3.11
protobuf==4.22.0
psycopg2-binary
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycryptodome==3.4.3
pyparsing==3.0.9
python-jwt==2.0.1
pytz==2022.7.1
requests==2.28.2
requests-toolbelt==0.7.0
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stone==3.3.1
tzdata==2022.7
urllib3==1.26.14
whitenoise==6.4.0
